// src/app/services/swagger/model/article.ts
/**
 * BoF REST API
 * Battle for honor interface
 *
 * OpenAPI spec version: 0.0.1
 * Contact: bla@bla.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * .
 */
export interface Article { 
    /**
     * The id
     */
    idArticle: number;
    /**
     * The title
     */
    title: string;
    /**
     * The language code
     */
    langCode: string;
    /**
     * The category
     */
    wikiCategory: Article.WikiCategoryEnum;
}
export namespace Article {
    export type WikiCategoryEnum = 'WELCOME_MESSAGE' | 'GAME_MECHANICS';
    export const WikiCategoryEnum = {
        WELCOMEMESSAGE: 'WELCOME_MESSAGE' as WikiCategoryEnum,
        GAMEMECHANICS: 'GAME_MECHANICS' as WikiCategoryEnum
    };
}