// src/app/services/swagger/model/articlePlainContent.ts
/**
 * BoF REST API
 * Battle for honor interface
 *
 * OpenAPI spec version: 0.0.1
 * Contact: bla@bla.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ArticleRevision } from './articleRevision';

/**
 * .
 */
export interface ArticlePlainContent { 
    revision: ArticleRevision;
    /**
     * The current revision
     */
    content: string;
}